+++ src/main/java/com/sdl/selenium/bootstrap/button/Upload.java
@@ -2,5 +2,8 @@ package com.sdl.selenium.bootstrap.button;
 
 public interface Upload {
 
+    @Deprecated
     boolean upload(String... filePath);
+
+    boolean upload(String filePath);
 }