+++ src/main/java/com/sdl/selenium/web/button/SelectFilesHandler.java
@@ -41,12 +41,24 @@ public class SelectFilesHandler {
 
     /**
      * Upload file with AutoIT exe
-     * Use only this: button.browseWithAutoIT(new String[] {"C:\\upload.exe", "C:\\text.txt"});
-     *
+     * Use only this: button.browseWithAutoIT("C:\\upload.exe", "C:\\text.txt");
+     * @deprecated The next version, replaced by {@link #browseWithAutoIT(String)}.
      * @param filePath   path to upload.exe
      * @return true or false
      */
-    public boolean browseWithAutoIT(String[] filePath) {
+    public boolean browseWithAutoIT(String... filePath) {
+        openBrowseWindow();
+        return RunExe.getInstance().upload(filePath);
+    }
+
+    /**
+     * Upload file with AutoIT exe
+     * Use only this: button.browseWithAutoIT("C:\\text.txt");
+     *
+     * @param filePath   path to file
+     * @return true or false
+     */
+    public boolean browseWithAutoIT(String filePath) {
         openBrowseWindow();
         return RunExe.getInstance().upload(filePath);
     }