+++ src/main/java/com/sdl/selenium/web/button/UploadFile.java
@@ -20,12 +20,54 @@ public class UploadFile extends WebLocator implements Upload {
         withContainer(container);
     }
 
+    /**
+     * Upload file with AutoIT.
+     * Use only this: button.upload(this, "C:\\upload.exe", "C:\\text.txt");
+     * @deprecated The next version, replaced by {@link #upload(String)}.
+     *
+     * @param filePath e.g. "C:\\upload.exe", "C:\\text.txt"
+     * @return true | false
+     */
+    @Deprecated
     @Override
     public boolean upload(String... filePath) {
         return upload(this, filePath);
     }
 
-    public boolean upload(WebLocator el, String... filePath) {
+    /**
+     * Upload file with AutoIT.
+     * Use only this: button.upload("C:\\text.txt");
+     *
+     * @param filePath e.g. "C:\\text.txt"
+     * @return true | false
+     */
+    @Override
+    public boolean upload(String filePath) {
+        return upload(this, filePath);
+    }
+
+    /**
+     * Upload file with AutoIT.
+     * Use only this: button.upload(this, "C:\\upload.exe", "C:\\text.txt");
+     * @deprecated The next version, replaced by {@link #upload(WebLocator, String)}.
+     * @param el the item that you click to open upload window
+     * @param filePath e.g. "C:\\upload.exe", "C:\\text.txt"
+     * @return true | false
+     */
+    @Deprecated
+    private boolean upload(WebLocator el, String... filePath) {
+        return executor.browse(el) && executor.upload(filePath);
+    }
+
+    /**
+     * Upload file with AutoIT.
+     * Use only this: button.upload(this, "C:\\text.txt");
+     *
+     * @param el the item that you click to open upload window
+     * @param filePath e.g. "C:\\text.txt"
+     * @return true | false
+     */
+    private boolean upload(WebLocator el, String filePath) {
         return executor.browse(el) && executor.upload(filePath);
     }
 }