+++ src/test/functional/java/com/sdl/selenium/bootstrap/button/UploadFileIntegrationTest.java
@@ -14,7 +14,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
 
 public class UploadFileIntegrationTest extends TestBase {
-    private static final Logger LOGGER = LoggerFactory.getLogger(UploadFileIntegrationTest.class);
 
     private Form form = new Form(null, "Form Title");
     private UploadFile uploadFile = new UploadFile(form, "TPT Test:");
@@ -26,19 +25,19 @@ public class UploadFileIntegrationTest extends TestBase {
 
     @Test
     public void upload() {
-        uploadFile.upload("Select file", new String[]{InputData.UPLOAD_EXE_PATH, InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\example.txt"});
+        uploadFile.upload("Select file", InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\example.txt");
         assertThat(new File(InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\example.txt").getName(), equalTo(uploadFile.uploadedNameFile()));
     }
 
     @Test
     public void uploadSimple() {
-        uploadFile.upload(InputData.UPLOAD_EXE_PATH, InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\example.txt");
+        uploadFile.upload(InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\example.txt");
         assertThat(new File(InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\example.txt").getName(), equalTo(uploadFile.uploadedNameFile()));
     }
 
     @Test(dependsOnMethods = "upload")
     public void change() {
-        uploadFile.change(new String[]{InputData.UPLOAD_EXE_PATH, InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\exampleM.txt"});
+        uploadFile.change(InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\exampleM.txt");
         assertThat(new File(InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\exampleM.txt").getName(), equalTo(uploadFile.uploadedNameFile()));
     }
 }