+++ release-notes.md
@@ -2,8 +2,11 @@ RELEASE NOTES
 -------------
 **Release Notes for Testy 2.2.0-SNAPSHOT**
 - change set***() in with***()
-  ex: WebLocator el = new WebLocator().setClasses("x-btn").setText("Save").setType("text");
-      WebLocator el = new WebLocator().withClasses("x-btn").withText("Save").withType("text");
+    ex: WebLocator el = new WebLocator().setClasses("x-btn").setText("Save").setType("text");
+        WebLocator el = new WebLocator().withClasses("x-btn").withText("Save").withType("text");
+-  Add possibility to set in browser config: upload.exe.path=src\\test\\resources\\upload\\upload.exe
+    ex: UploadFile uploadButton = new UploadFile().withText("Browse");
+        uploadButton.upload(InputData.RESOURCES_DIRECTORY_PATH + "\\upload\\text.docx");
 
 **Release Notes for Testy 2.1.0**
 - use attributes when generate toString, in case no other important string is present